work-in-progress react native app implemented
currently, the app is capable of uploading videos and has an analysis tab to show the videos with some (placeholder) analysis data. we need to fine tune the UI, connect the app to the actual analyzer backend, and fix the non-functioning parts of the react app (uploading more than one video, aspect ratio of videos appearing off, etc.) So far the app has only been tested on web.